+++ src/gameData.ts
@@ -1,0 +1,6 @@
+import { ScoreTableItem } from "./entities/ScoreTableItem.type";
+
+export let games = new Map();
+export let players = new Map();
+export let rooms = new Map();
+export let scoreTable: ScoreTableItem[] = [];
